i18n rules initialization for MySQL
